+++ plugins/json-schema/src/index.ts
@@ -1,0 +1,19 @@
+import { Flatfile } from '@flatfile/api'
+import { FlatfileEvent, FlatfileListener } from '@flatfile/listener'
+import { configureSpace } from '@flatfile/plugin-space-configure'
+import { JsonSetupFactory, generateSetup } from './setup.factory'
+
+export function configureSpaceWithJsonSchema(
+  setupFactory: JsonSetupFactory,
+  callback?: (
+    event: FlatfileEvent,
+    workbookIds: string[],
+    tick: (progress?: number, message?: string) => Promise<Flatfile.JobResponse>
+  ) => any | Promise<any>
+) {
+  return async function (listener: FlatfileListener) {
+    listener.use(configureSpace(await generateSetup(setupFactory), callback))
+  }
+}
+
+export * from './setup.factory'