+++ plugins/record-hook/src/tests/bulkRecordHook.simpleData.e2e.spec.ts
@@ -113,4 +113,33 @@ describe('bulkRecordHook() simple data modification e2e', () => {
       expect(errors.length).toBe(records.length)
     })
   })
+
+  describe('bulkRecordHook errors', () => {
+    const logErrorSpy = jest.spyOn(global.console, 'error')
+    beforeEach(async () => {
+      listener.use(
+        bulkRecordHook('test', (records) =>
+          records.map((record) => {
+            throw new Error('oops')
+          })
+        )
+      )
+      await createRecords(sheetId, [
+        {
+          name: 'John Doe',
+          email: 'john@doe.com',
+          notes: 'foobar',
+        },
+      ])
+    })
+    it('returns readable error', async () => {
+      await listener.waitFor('commit:created')
+      expect(logErrorSpy).toHaveBeenCalledWith(
+        'An error occurred while running the handler: oops'
+      )
+    })
+    afterEach(() => {
+      logErrorSpy.mockClear()
+    })
+  })
 })