+++ plugins/record-hook/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @flatfile/plugin-record-hook
 
+## 1.1.12
+
+### Patch Changes
+
+- 28820d5: Release to update @flatfile/api dep
+- Updated dependencies [28820d5]
+  - @flatfile/util-common@0.2.3
+
+## 1.1.11
+
+### Patch Changes
+
+- 4d3fd49: Bug fix for record comparison
+
 ## 1.1.10
 
 ### Patch Changes